ml/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
fastapi==0.115.6
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
